+++ bricks/stx_project_template/__brick__/lib/router/router_observer.dart
@@ -1,6 +1,5 @@
-import 'package:flutter/material.dart';
-
 import 'package:auto_route/auto_route.dart';
+import 'package:material_ui/material_ui.dart';
 
 import 'package:{{project_name}}/services/logger/index.dart';