+++ bricks/stx_project_template/__brick__/lib/widgets/form_builder/text_input_form_builder.dart
@@ -1,8 +1,7 @@
 import 'dart:math';
 
-import 'package:flutter/material.dart';
-
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:stx_flutter_form_bloc/stx_flutter_form_bloc.dart';
 
 import 'package:{{project_name}}/theme/index.dart';