+++ lib/app/app_initialization.dart
@@ -1,10 +1,9 @@
 import 'dart:ui';
 
-import 'package:flutter/material.dart';
-
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
+import 'package:material_ui/material_ui.dart';
 import 'package:stx_flutter_form_bloc/stx_flutter_form_bloc.dart';
 
 import 'package:flutter_guidelines/blocs/bloc_observer.dart';